refactor(components): add modular index file to all components folder
