+++ packages/components/src/components/button/Chip.tsx
@@ -4,7 +4,7 @@ import {StyleSheet} from 'react-native'
 import {useBlossomTheme} from '../../context'
 import {ChipProps} from '../types'
 import {useMergedProps} from '../../common'
-import Icon from '../icon'
+import {Icon} from '../icon'
 import BaseButton from './BaseButton'
 import {View} from '../view'
 import {getTextColorName} from '../utils'
