+++ packages/components/src/components/button/index.ts
@@ -1,0 +1,2 @@
+export {default as Chip} from './Chip'
+export {default as Button} from './Button'
