+++ packages/components/src/components/icon/index.ts
@@ -1,0 +1,1 @@
+export {default as Icon} from './Icon'
