+++ packages/components/src/components/image/index.ts
@@ -1,0 +1,1 @@
+export {default as Avatar} from './Avatar'
