+++ packages/components/src/components/index.ts
@@ -1,15 +1,12 @@
-export {default as Text} from './text/Text'
-export {default as Button} from './button'
-export {default as Chip} from './button/Chip'
-export {default as ActivityIndicator} from './loader/ActivityIndicator'
-export {default as TextInput} from './input/TextInput'
-export {default as SearchInput} from './input/SearchInput'
-export {default as Avatar} from './image/Avatar'
+export * from './text'
+export * from './loader'
+export * from './input'
+export * from './image'
 export * from './formFields'
 export * from './view'
+export * from './button'
 export * from './modal'
 export * from './select'
-
-export {default as Icon} from './icon'
+export * from './icon'
 
 export * from './types'
