+++ packages/components/src/components/input/index.ts
@@ -1,0 +1,2 @@
+export {default as SearchInput} from './SearchInput'
+export {default as TextInput} from './TextInput'
