+++ packages/components/src/components/loader/index.ts
@@ -1,0 +1,1 @@
+export {default as ActivityIndicator} from './ActivityIndicator'
