+++ packages/components/src/components/text/index.ts
@@ -1,3 +1,1 @@
-export {typographyStyle} from './typography'
-export {default as SizedText} from './SizedText'
 export {default as Text} from './Text'
